+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Version 1.135-dev (not yet released
 -------------
 
 * Removed check on related deleted entities. See #43
+* Added action: Participant: Calculate Participant Fee
+* Added action: Participant: Create contribution for participant
 
 Version 1.134
 -------------